fileSD: Do not try to remove non-existent lease file
The volume.lease file exists only if a domain has volume leases. Avoid
removing the non-existent lease file and the bogus exception in the log
when the operation "fails".

Bug-Url: https://bugzilla.redhat.com/1291207
Reviewed-on: https://gerrit.ovirt.org/50589
Continuous-Integration: Jenkins CI

+++ vdsm/storage/fileSD.py
@@ -219,9 +219,10 @@ class FileStorageDomainManifest(sd.StorageDomainManifest):
                 metaFile = volPath + '.meta'
                 self.log.debug("Removing file: %s", metaFile)
                 self.oop.os.remove(metaFile)
-                leaseFile = volPath + '.lease'
-                self.log.debug("Removing file: %s", leaseFile)
-                self.oop.os.remove(leaseFile)
+                if self.hasVolumeLeases():
+                    leaseFile = volPath + '.lease'
+                    self.log.debug("Removing file: %s", leaseFile)
+                    self.oop.os.remove(leaseFile)
             except OSError:
                 self.log.error("vol: %s can't be removed.",
                                volPath, exc_info=True)
